+++ src/pytorch_lightning/serve/__init__.py
@@ -1,0 +1,4 @@
+from pytorch_lightning.serve.module import ServableModule
+from pytorch_lightning.serve.sanity_serving import SanityServing
+
+__all__ = ["SanityServing", "ServableModule"]